RavenDB-14274 - fix upload progress
