+++ src/Raven.Server/Documents/PeriodicBackup/Progress.cs
@@ -1,5 +1,8 @@
 using System;
+using System.Diagnostics;
 using Raven.Client.Documents.Operations.Backups;
+using Raven.Server.Utils.Metrics;
+using Sparrow;
 
 namespace Raven.Server.Documents.PeriodicBackup
 {
@@ -13,5 +16,32 @@ namespace Raven.Server.Documents.PeriodicBackup
         public UploadProgress UploadProgress { get; }
 
         public Action OnUploadProgress { get; set; } = () => { };
+
+        public static Progress Get(UploadProgress uploadProgress, Action<string> onProgress)
+        {
+            var bytesPutsPerSec = new MeterMetric();
+
+            long lastUploadedInBytes = 0;
+            var sw = Stopwatch.StartNew();
+            var progress = new Progress(uploadProgress)
+            {
+                OnUploadProgress = () =>
+                {
+                    if (sw.ElapsedMilliseconds <= 2000)
+                        return;
+
+                    var totalUploadedInBytes = uploadProgress.UploadedInBytes;
+                    bytesPutsPerSec.MarkSingleThreaded(totalUploadedInBytes - lastUploadedInBytes);
+                    lastUploadedInBytes = totalUploadedInBytes;
+                    var uploaded = new Size(totalUploadedInBytes, SizeUnit.Bytes);
+                    var totalToUpload = new Size(uploadProgress.TotalInBytes, SizeUnit.Bytes);
+                    uploadProgress.BytesPutsPerSec = bytesPutsPerSec.MeanRate;
+                    onProgress($"Uploaded: {uploaded} / {totalToUpload}");
+                    sw.Restart();
+                }
+            };
+
+            return progress;
+        }
     }
 }
